refactor: :recycle: Move `run` function into separate module
