+++ src/gpu/pdb_gpu/main.rs
@@ -1,4 +1,4 @@
-use pdb_tui::gpu::pdb_gpu::run;
+use pdb_tui::gpu::pdb_gpu::run::run;
 
 fn main() {
     pollster::block_on(run());
